requirements.txt
streamlit
pandas
scikit-learn
numpy
google-generativeai